Fix summary about models (#8245)
* Expose and document utils.layer_utils.print_summary as utils.print_summary

Fixes #8105

* Fix docs autogen when running on python 3.6.2

When a class inherits from object without defining __init__,
AttributeError is raised instead of TypeError. So we need to catch
both exceptions to be compatible.

* Fix process_XXX_docstring functions to work with models function

The process_function_docstring would fail to parse the “# Argument”
and “# Returns” of class functions. This was because it expected the
“# Argument” string to be preceeded by exactly a newline and 4 spaces,
but for class methods, that would be a newline and 8 spaces.

This fixes the problem by expecting a newline and some number of space.